refactor(comfyui): 進捗の記録の読み書きを`manifest.py`へ分けます
`read_manifest`が壊れた記録をどう扱うかは、
実際に中断したジョブでしか通らない経路です。
壊れると「再開したつもりが生成条件の照合を飛ばす」、
「完了しているのに記録の上では未完了に見える」といった、
気付きにくい失敗になります。

しかし`__init__.py`はComfyUI本体とtorchとPyAVをimportするので、
読み込むだけでGPUを持つ実行環境が要り、
テストから触れませんでした。

記録の読み書きはファイルとJSONだけで完結しているので、
`Segment`と`TranslatedSegment`と`Manifest`、
`optional_count`と`read_manifest`と`write_manifest`を`manifest.py`へ移して、
素のPythonで読み込める形にします。

共有モジュールと同じくノードのディレクトリ内に置くので、
ComfyUIが読む配置も`cp -rL`で配る仕組みも変わりません。
移動だけで処理は変えていません。

型検査とbulletのビルドが通ることを確認しました。
